66239 - Fix Partner required on Move when we select Purchase or Sale journal
