+++ axelor-account/src/main/java/com/axelor/apps/account/service/move/attributes/MoveAttrsService.java
@@ -65,4 +65,6 @@ public interface MoveAttrsService {
   void addMassEntryPaymentConditionRequired(Move move, Map<String, Map<String, Object>> attrsMap);
 
   void addMassEntryBtnHidden(Move move, Map<String, Map<String, Object>> attrsMap);
+
+  void addPartnerRequired(Move move, Map<String, Map<String, Object>> attrsMap);
 }
