+++ axelor-account/src/main/java/com/axelor/apps/account/service/moveline/MoveLineAttrsService.java
@@ -59,7 +59,5 @@ public interface MoveLineAttrsService {
 
   void addAnalyticDistributionTemplateDomain(Move move, Map<String, Map<String, Object>> attrsMap);
 
-  void addPartnerRequired(Move move, Map<String, Map<String, Object>> attrsMap);
-
   void changeFocus(Move move, MoveLine moveLine, Map<String, Map<String, Object>> attrsMap);
 }
