+++ axelor-account/src/main/java/com/axelor/apps/account/service/moveline/MoveLineAttrsServiceImpl.java
@@ -20,7 +20,6 @@ package com.axelor.apps.account.service.moveline;
 
 import com.axelor.apps.account.db.*;
 import com.axelor.apps.account.db.repo.AccountRepository;
-import com.axelor.apps.account.db.repo.JournalTypeRepository;
 import com.axelor.apps.account.db.repo.MoveRepository;
 import com.axelor.apps.account.service.PeriodServiceAccount;
 import com.axelor.apps.account.service.analytic.AnalyticLineService;
@@ -290,20 +289,6 @@ public class MoveLineAttrsServiceImpl implements MoveLineAttrsService {
             "self.isSpecific IS FALSE AND self.company.id = %d", move.getCompany().getId());
 
     this.addAttr("analyticDistributionTemplate", "domain", domain, attrsMap);
-  }
-
-  @Override
-  public void addPartnerRequired(Move move, Map<String, Map<String, Object>> attrsMap) {
-    Objects.requireNonNull(move);
-    boolean required =
-        move.getJournal() != null
-                && move.getJournal().getJournalType() != null
-                && move.getJournal().getJournalType().getTechnicalTypeSelect()
-                    == JournalTypeRepository.TECHNICAL_TYPE_SELECT_EXPENSE
-            || move.getJournal().getJournalType().getTechnicalTypeSelect()
-                == JournalTypeRepository.TECHNICAL_TYPE_SELECT_SALE;
-
-    this.addAttr("partner", "required", required, attrsMap);
   }
 
   @Override
