+++ axelor-budget/src/main/java/com/axelor/apps/budget/service/move/MoveLineGroupBudgetServiceImpl.java
@@ -5,6 +5,7 @@ import com.axelor.apps.account.db.MoveLine;
 import com.axelor.apps.account.service.analytic.AnalyticLineService;
 import com.axelor.apps.account.service.move.MoveLineInvoiceTermService;
 import com.axelor.apps.account.service.move.MoveToolService;
+import com.axelor.apps.account.service.move.attributes.MoveAttrsService;
 import com.axelor.apps.account.service.moveline.MoveLineAttrsService;
 import com.axelor.apps.account.service.moveline.MoveLineCheckService;
 import com.axelor.apps.account.service.moveline.MoveLineComputeAnalyticService;
@@ -37,6 +38,7 @@ public class MoveLineGroupBudgetServiceImpl extends MoveLineGroupBankPaymentServ
       MoveLineToolService moveLineToolService,
       MoveToolService moveToolService,
       AnalyticLineService analyticLineService,
+      MoveAttrsService moveAttrsService,
       MoveLineCheckBankPaymentService moveLineCheckBankPaymentService,
       MoveLineRecordBankPaymentService moveLineRecordBankPaymentService,
       BudgetToolsService budgetToolsService) {
@@ -51,6 +53,7 @@ public class MoveLineGroupBudgetServiceImpl extends MoveLineGroupBankPaymentServ
         moveLineToolService,
         moveToolService,
         analyticLineService,
+        moveAttrsService,
         moveLineCheckBankPaymentService,
         moveLineRecordBankPaymentService);
     this.budgetToolsService = budgetToolsService;
